Remove concrete `Load` type
We shouldn't force the representation of load. This is something that
really should be determined by the implementai of `Choose`.

The `Loaded` trait is now called `Load` and has a type parameter,
`Metric`, that determines what load metric type is required.

load::Constant is generic over its load metric type.

load::PendingRequests now simply exposes the total count of pending
requests, without having to coerce to a float on [0.0, 1.0].

choose::RoundRobin uses a unit-typed load metric to indicate that it
ignores load.

choose::PowerOfTwoChoices only requires that the metric be ordered.