Support weighted balancing
In order to implement red-line testing, blue-green deployments, and
other operational use cases, many service discovery and routing schemes
support weighting.

In this iteration, we provide a decorator type, `Weighted`, that may be
used to wrap load-bearing services to alter their load according to a
weight.  The `Weighted` type may also be used to wrap `Discover`
implementations, in which case it will wrap all new services with
`Weighted`.