tower-balance: Handle duplicate Insert events by overwriting (#75)
When an endpoint's state changes in some way, it may need to be rebound to a 
new service, and reinserted into the load balancer. This PR changes 
`tower-balance` so that, rather than ignoring duplicate `Insert`s, the new
endpoint replaces the old endpoint. The new endpoint is always placed on the
not-ready list; if the replaced endpoint was on the ready list, it is removed
prior to inserting the new endpoint into the not-ready list.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>