balance: Account for long-lived responses (#74)
tower-balance provides a PendingRequests load metric that counts the number of responses
that have not yet been received. However, especially in the case of HTTP, responses have
bodies that remain active far past the initial receipt of the response. We want load
metrics to be able to take such streams into account.

This change introduces a new utility trait, `load::Measure`, which is used by implementors
of `Load` (like `PendingRequests`) to handle the protocol-specific details of attaching an
_instrument_ with a response message.  _Instruments_ are implemented as RAII-guarded types
that ensure that load calculations are updated as a response completes.  An instrument is
dropped when the load metric no longer needs information from a response.

This all being changed in service of a `PeakEwma` balancer implementation, though this
should benefit the existing load metric as well.