Don't optimize poll_service -> Ready in Balance
Previously there was an optimization that tracked which services in the
`Balance` service set had previously returned `Ready` for
`poll_service`, and avoided calling `poll_service` on them again.
However, it was incomplete; it did not add services back to the set when
`call` was later invoked. Doing so would either require `D::Key: Clone`,
or some hairy tracking to use the indices of the `IndexMap` (which
change as services move between `ready` and `not_ready`). This patch
instead removes the optimization entirely, and leaves it as a TODO for
later.