Add CallAll combinator.
This PR adds the `CallAll` combinator, which essentially provides

```rust
Stream<Item = Request> + Service<Request> => Stream<Item = Response>
```

That is, given a `Stream` of requests and a `Service` that can accept
those requests, it produces a `Stream` of the corresponding responses.

To allow CallAll to return the wrapped `Service` when there are no more
requests, CallAll is _really_ more like a
[`StateStream`](https://docs.rs/futures-state-stream/), so we may want
to consider pulling that in for compatibility. For the time being
though, this is simply exposed through a `StateStreamItem` enum.