Add basic "batteries-included" `retry::Policy`s.
These are lifted out of the test and example code.  Two policies are provided:

* `RetryLimit`, with a bounded number of retry attempts;
* `RetryError`, with unbounded retry attempts.

Both policies require `Req: Clone` as it's not possible to write a generic
retry policy without being able to clone requests.

The docs should be updated with a little blurb that explains when they should
not be used.  Also, the `Policy` docs had an `Attempts` example corresponding
to `RetryLimit` (formerly `Limit` in the tests code); I removed it because
`RetryLimit` exists and can be view-sourced.