util: Add `ServiceExt::map_future` (#542)
* util: Add `ServiceExt::map_future`

I ran into a thing today where I wanted to write a middleware that wraps
all futures produced by a service in a `tracing::Span`. So something
like `self.inner.call(req).instrument(span)`.

Afaik all the combinators we have today receive the value produced by
the future and not the future itself. At that point its too late to call
`.instrument`. So I thought this made sense to add a combinator for.

* Add additional trait bounds to improve UX

* Add docs

* Update changelog

* Clean up debug impl

* Better debug impl for `MapFutureLayer`