buffer: add test for `poll_ready` leaking permits
This reproduces an issue in previous versions of `tower::buffer`, where
multiple calls to `poll_ready` when the buffer is already ready would
acquire additional channel capacity that was never used. This
essentially leaked channel capacity.

This issue doesn't exist with the current implementation, but this test
should help ensure it isn't reintroduced when changing the buffer
internals (such as switching to `tokio-util`'s pollable channel).

Signed-off-by: Eliza Weisman <eliza@buoyant.io>