ci: disable fail-fast for Rust version matrix jobs
Tower has test and check jobs on CI that run on a build matrix
including a number of Rust versions. By default, GitHub Actions has
fail-fast semantics for matrix jobs, so if any matrix job fails, the
rest are cancelled and the build is failed. This is intended to help
builds complete faster.

This isn't really the ideal behavior for testing across multiple Rust
versions. When a build fails on a particular toolchain version, we would
ideally like to know whether the failure is localized to that version or
exists on _all_ Rust versions. This is particularly important for builds
on nightly Rust, as the nightly toolchain is more likely to contain
compiler regressions that might not be our fault at all. Similarly, we
might want to know if a change only broke the build on our MSRV, or if
it broke the build everywhere --- such an issue would be fixed
differently.

Therefore, this PR changes the CI workflow to disable fail-fast behavior
on the cross-version test jobs. This way, we will always see the results
of the jobs for each Rust toolchain version. This does *not* change what
jobs are required to merge a PR --- all versions must still pass (which
we may also want to change, but that's a separate issue).