rewrite *again* to use `PollSender`
in `tokio-util` v0.7, the semantics of `PollSender` was changed to have
a `poll_reserve` rather than `poll_send_done`, which is the required
behavior for `Buffer`. The implementation of `PollSender` now looks more
or less identical to the code I had written using `ReusableBoxFuture`
and `reserve_owned`. Therefore, we can now just use the
`tokio_util::sync::PollSender` type, and the buffer implementation is
now much simpler.